change the creation of the test user, employee and role to a command. Use '$ flask createdb' before first run in development mode.
The App will be created now with a application Factory method cause of creating multiple instances of the app.

Some minor changes to the config file because the ability to create multiple instances of the app
